refactor(catalog): create folder and separate file for all components
