+++ src/components/Logo/Logo.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function Logo() {
+	return (
+		<div className="login-page__logo">
+			<div className="login-page__logo-item"></div>
+		</div>
+	);
+}
